refactor: organize book model code (Jose, Valery, Cesar)
