+++ frontend/public/data/quices.js
@@ -1,0 +1,92 @@
+const questions= [
+    {
+        section: 1,
+        questions: [
+          {
+            question: "¿Cuál es el dios del Rayo?",
+            options: [
+              { option: "Zeus", correct: true },
+              { option: "Hades", correct: false },
+              { option: "cronos", correct: false },
+              { option: "Barranquilla", correct: false }
+            ]
+
+          },
+          {
+            question: "¿Cuál es el dios del Rayo?",
+            options: [
+              { option: "Zeus", correct: true },
+              { option: "Hades", correct: false },
+              { option: "cronos", correct: false },
+              { option: "Barranquilla", correct: false }
+            ]
+
+          },
+        ]
+
+    },
+    {
+        section: 2,
+        questions: [
+          {
+            question: "¿Cuál es la capital de Colombia?",
+            options: [
+              { option: "Bogotá", correct: true },
+              { option: "Medellín", correct: false },
+              { option: "Cali", correct: false },
+              { option: "Barranquilla", correct: false }
+            ]
+
+          },
+        ]
+
+    },
+    {
+        section: 3,
+        questions: [
+          {
+            question: "¿Cuál es la capital de Colombia?",
+            options: [
+              { option: "Bogotá", correct: true },
+              { option: "Medellín", correct: false },
+              { option: "Cali", correct: false },
+              { option: "Barranquilla", correct: false }
+            ]
+
+          },
+        ]
+
+    },
+    {
+        section: 4,
+        questions: [
+          {
+            question: "¿Cuál es la capital de Colombia?",
+            options: [
+              { option: "Bogotá", correct: true },
+              { option: "Medellín", correct: false },
+              { option: "Cali", correct: false },
+              { option: "Barranquilla", correct: false }
+            ]
+
+          },
+        ]
+
+    },
+    {
+        section: 5,
+        questions: [
+          {
+            question: "¿Cuál es la capital de Colombia?",
+            options: [
+              { option: "Bogotá", correct: true },
+              { option: "Medellín", correct: false },
+              { option: "Cali", correct: false },
+              { option: "Barranquilla", correct: false }
+            ]
+
+          },
+        ]
+
+    },
+]
