+++ frontend/public/data/stickers.js
@@ -1,13 +1,11 @@
 const stickers = [
   {
-    // id: 1,
     page: 1,
     texture: '/img/world/zeus.jpeg',
     text: 'Zeus, dios supremo del cielo, el trueno y el rayo,  padre de los dioses y los hombres.  Posee gran poder y es representado como un hombre barbudo con una poderosa figura.',
     position: [0.8, 0.5, -0.25],
   },
   {
-    // id: 2,
     page: 1,
     texture: '/img/world/zeus.jpeg',
     text: 'XD',
